Create booking invitation

+++ src/bookings/bookings.controller.ts
@@ -129,14 +129,21 @@ export class BookingsController {
     }
   }
 
-  // Allows admin update all one-time bookings
-  @Roles(Role.ADMIN)
-  @Put('one-time/admin/:id')
-  async updateOneTimeBookingAdmin(
+  // Allows user updates only own one-time bookings and admin all one-time bookings
+  @Put('one-time/:id')
+  async updateOneTimeBooking(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateOneTimeBookingDto: UpdateOneTimeBookingDto,
+    @Request() req,
   ): Promise<OneTimeBooking> {
     try {
+      if (req.user.role !== Role.ADMIN) {
+        return await this.oneTimeBookingsService.updateOwn(
+          id,
+          req.user.id,
+          updateOneTimeBookingDto,
+        );
+      }
       return await this.oneTimeBookingsService.update(
         id,
         updateOneTimeBookingDto,
@@ -150,72 +157,21 @@ export class BookingsController {
     }
   }
 
-  // Allows user update only own one-time bookings
-  @Roles(Role.USER)
-  @Put('one-time/user/:id')
-  async updateOneTimeBookingUser(
-    @Param('id', ParseIntPipe) id: number,
-    @Body() updateOneTimeBookingDto: UpdateOneTimeBookingDto,
-    @Request() req,
-  ): Promise<OneTimeBooking> {
-    try {
-      return await this.oneTimeBookingsService.updateOwn(
-        id,
-        req.user.id,
-        updateOneTimeBookingDto,
-      );
-    } catch (error) {
-      if (error instanceof ServiceException) {
-        throw new HttpException(error.message, 404);
-      } else {
-        throw error;
-      }
-    }
-  }
-
-  // Allows admin delete all one-time bookings
-  @Roles(Role.ADMIN)
-  @Delete('one-time/admin/:id')
-  async deleteOneTimeBookingAdmin(
-    @Param('id', ParseIntPipe) id: number,
-  ): Promise<DeleteResult> {
-    try {
-      return await this.oneTimeBookingsService.delete(id);
-    } catch (error) {
-      if (error instanceof ServiceException) {
-        throw new HttpException(error.message, 404);
-      } else {
-        throw error;
-      }
-    }
-  }
-
-  // Allows user delete only own one-time bookings
-  @Roles(Role.USER)
-  @Delete('one-time/user/:id')
-  async deleteOneTimeBookingUser(
-    @Param('id', ParseIntPipe) id: number,
-    @Request() req,
-  ): Promise<DeleteResult> {
-    try {
-      return await this.oneTimeBookingsService.deleteOwn(id, req.user.id);
-    } catch (error) {
-      if (error instanceof ServiceException) {
-        throw new HttpException(error.message, 404);
-      } else {
-        throw error;
-      }
-    }
-  }
-
-  // Allows admin update all recurring bookings
-  @Roles(Role.ADMIN)
-  @Put('recurring/admin/:id')
+  // Allows user updates only own recurring bookings and admin all recurring bookings
+  @Put('recurring/:id')
   async updateRecurringBookingAdmin(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateRecurringBookingDto: UpdateRecurringBookingDto,
+    @Request() req,
   ): Promise<RecurringBooking> {
     try {
+      if (req.user.role !== Role.ADMIN) {
+        return await this.recurringBookingsService.updateOwn(
+          id,
+          req.user.id,
+          updateRecurringBookingDto,
+        );
+      }
       return await this.recurringBookingsService.update(
         id,
         updateRecurringBookingDto,
@@ -229,20 +185,17 @@ export class BookingsController {
     }
   }
 
-  // Allows user update only own recurring bookings
-  @Roles(Role.USER)
-  @Put('recurring/user/:id')
-  async updateRecurringBookingUser(
+  // Allows user deletes only own one-time bookings and admin all one-time bookings
+  @Delete('one-time/:id')
+  async deleteOneTimeBooking(
     @Param('id', ParseIntPipe) id: number,
-    @Body() updateRecurringBookingDto: UpdateRecurringBookingDto,
     @Request() req,
-  ): Promise<RecurringBooking> {
+  ): Promise<DeleteResult> {
     try {
-      return await this.recurringBookingsService.updateOwn(
-        id,
-        req.user.id,
-        updateRecurringBookingDto,
-      );
+      if (req.user.role !== Role.ADMIN) {
+        return await this.oneTimeBookingsService.deleteOwn(id, req.user.id);
+      }
+      return await this.oneTimeBookingsService.delete(id);
     } catch (error) {
       if (error instanceof ServiceException) {
         throw new HttpException(error.message, 404);
@@ -252,32 +205,17 @@ export class BookingsController {
     }
   }
 
-  // Allows admin delete all recurring bookings
-  @Roles(Role.ADMIN)
-  @Delete('recurring/admin/:id')
-  async deleteRecurringBookingAdmin(
+  // Allows user deletes only own recurring bookings and admin all recurring bookings
+  @Delete('recurring/:id')
+  async deleteRecurringBooking(
     @Param('id', ParseIntPipe) id: number,
+    @Request() req,
   ): Promise<DeleteResult> {
     try {
+      if (req.user.role !== Role.ADMIN) {
+        return await this.recurringBookingsService.deleteOwn(id, req.user.id);
+      }
       return await this.recurringBookingsService.delete(id);
-    } catch (error) {
-      if (error instanceof ServiceException) {
-        throw new HttpException(error.message, 404);
-      } else {
-        throw error;
-      }
-    }
-  }
-
-  // Allows user delete only own one-time bookings
-  @Roles(Role.USER)
-  @Delete('recurring/user/:id')
-  async deleteRecurringBookingUser(
-    @Param('id', ParseIntPipe) id: number,
-    @Request() req,
-  ): Promise<DeleteResult> {
-    try {
-      return await this.recurringBookingsService.deleteOwn(id, req.user.id);
     } catch (error) {
       if (error instanceof ServiceException) {
         throw new HttpException(error.message, 404);
